Some fixes and optimizations Implemented profiling for LSHTC

+++ src/c/parameter.h
@@ -5,14 +5,14 @@ typedef struct
 {
   double C1;   // the penalty for an example being outside it's class bounary
   double C2;   // the penalty for an example being inside other class' boundary
-  unsigned int max_iter;  // maximum number of iterations
+  unsigned long max_iter;  // maximum number of iterations
   int batch_size; // size of the minibatch
   double eps; // not used
   double eta; // the initial learning rate. The leraning rate is eta/sqrt(t) where t is the number of iterations
   double min_eta; //the minimum value of the lerarning rate (i.e. lr will be max (eta/sqrt(t), min_eta)
-  int reorder_epoch; //number of iterations between class reorderings. 0 for no reordering of classes
+  unsigned long reorder_epoch; //number of iterations between class reorderings. 0 for no reordering of classes
   int max_reorder; // maxumum number of class reorderings. Each reordering runs until convergence or for Max_Iter iterations and after each reordering the learning rate is reset
-  int report_epoch; //number of iterations between computation and report the objective value (can be expensive because obj is calculated on the entire training set). 0 for no reporting
+  long int report_epoch; //number of iterations between computation and report the objective value (can be expensive because obj is calculated on the entire training set). 0 for no reporting
   bool remove_constraints; // whether to remove the constraints for instances that fall outside the class boundaries in previous projections. 
   bool remove_class_constraints; // whether to remove the constraints for examples that fell outside their own class boundaries in previous projections. 
   bool rank_by_mean; // whether to rank the classes by the mean of the projected examples or by the midpoint of its [l,u] interval (i.e. (u-l)/2).
